Fix Engine State Debug Viewer
